+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# Version 0.15.1
+# Version 0.16.0
 
 * Add logo to Flowman Shell
 * Fix name of config option `flowman.execution.executor.class`
 * Add new `groupedAggregate` mapping
 * Reimplement target ordering, configurable via `flowman.execution.scheduler.class`
+* Implement new assertions `columns` and `expression`
 
 
 # Version 0.15.0 - 2021-03-23